# backend/ecollecto-backend/openapi.yaml
openapi: 3.1.0
info:
  title: eCollecto API
  description: eCollecto philatelic collection REST API. Public catalog endpoints
    are accessible without authentication. Protected endpoints require a valid Bearer
    JWT issued by Keycloak.
  version: v1
servers:
- url: http://localhost:8080
  description: Local development server
...
paths:
  /api/me/wishlist/items:
    post:
      tags:
      - Wishlist
      summary: Add to wishlist
      description: Adds a stamp to the authenticated user's wishlist.
      operationId: addToWishlist
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddWishlistItemRequest"
        required: true
      responses:
        "201":
          description: Stamp added to wishlist.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WishlistItemDto"
        "400":
          description: Invalid request body.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid JWT token.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Stamp already on wishlist.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - bearerAuth: []
...
  /api/me/wishlist:
    get:
      tags:
      - Wishlist
      summary: Get wishlist
      description: Returns all stamps on the authenticated user's wishlist.
      operationId: getWishlist
      responses:
        "200":
          description: Wishlist retrieved.
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItemDto"
        "401":
          description: Missing or invalid JWT token.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - bearerAuth: []
...
  /api/me/wishlist/items/{stampId}:
    delete:
      tags:
      - Wishlist
      summary: Remove from wishlist
      description: Removes a stamp from the authenticated user's wishlist.
      operationId: removeFromWishlist
      parameters:
      - name: stampId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Stamp removed from wishlist.
        "401":
          description: Missing or invalid JWT token.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Stamp not on wishlist.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
    ErrorResponse:
      type: object
      description: Error response payload.
      properties:
        message:
          type: string
          description: Error message.
        code:
          type: string
          description: Error code.
        status:
          type: integer
          format: int32
          description: HTTP status code.
    AddWishlistItemRequest:
      type: object
      description: Request body to add a stamp to the user's wishlist.
      properties:
        stampId:
          type: string
          description: Stamp ID to add.
          example: stamp-2024-001
          minLength: 1
      required:
      - stampId
    WishlistItemDto:
      type: object
      description: A stamp on the user's wishlist.
      properties:
        stampId:
          type: string
          description: Stamp ID (references stamps._id).
          example: stamp-2024-001
        addedAt:
          type: string
          format: date-time
          description: Timestamp when the stamp was added to the wishlist.
...
  securitySchemes:
    openIdConnect:
      type: openIdConnect
      description: OpenID Connect discovery for the eCollecto Keycloak realm. Use
        the ecollecto-ui public client with PKCE (S256) for the browser flow.
      openIdConnectUrl: http://localhost:8180/realms/ecollecto/.well-known/openid-configuration
    bearerAuth:
      type: http
      description: "JWT Bearer token issued by Keycloak (Authorization Code + PKCE).\
        \ Obtain a token from: http://localhost:8180/realms/ecollecto/protocol/openid-connect/token"
      scheme: bearer
      bearerFormat: JWT